Updated view GUI to match current version, renamed original view to legacy view
