+++ view.py
@@ -1,359 +1,505 @@
-
 import sys
-import controller
 from controller import *
-from PyQt5.QtWidgets import QApplication, QLabel, QPushButton, QVBoxLayout, QWidget, QFileDialog, QGridLayout,\
-    QTableWidget, QTableWidgetItem, QLineEdit, QComboBox
-from PyQt5 import QtGui, QtCore
-from PyQt5.QtGui import QCursor
+from PyQt5.QtGui import *
+from PyQt5.QtCore import *
+from PyQt5.QtWidgets import *
 
 
-widgets = {
-    "button_inventory": [],
-    "transaction": [],
-    "customer":[]
-}
+class HomepageWindow(QWidget):
+    def __init__(self):
+        super().__init__()
+        self.cams = None
+        self.title = "App"
+        self.top = 100
+        self.left = 100
+        self.width = 680
+        self.height = 500
+        self.initialize_windows()
 
-app = QApplication(sys.argv)
+    def initialize_windows(self):
+        self.setWindowTitle(self.title)
+        self.setGeometry(self.top, self.left, self.width, self.height)
+        self.setWindowIcon(self.style().standardIcon(QStyle.SP_DirClosedIcon))
 
-window = QWidget()
-window.setWindowTitle("Homepage")
-window.setFixedWidth(1200)
-window.setFixedHeight(800)
+        btn_view_products = QPushButton('View Products', self)
+        btn_view_products.clicked.connect(self.btn_view_products_on_click)
 
+        btn_view_oos_products = QPushButton('View Out of Stock Products', self)
+        btn_view_oos_products.clicked.connect(self.btn_view_out_of_stock_products_on_click)
 
-window.setStyleSheet("background:black;")
+        btn_view_recent_customers = QPushButton('View Customers Active in the Last Month', self)
+        btn_view_recent_customers.clicked.connect(self.btn_view_recent_customers_on_click)
 
-invoice_window = QWidget()
-invoice_window.setFixedWidth(1000)
-invoice_window.setFixedHeight(500)
-invoice_window.setStyleSheet("background:black;")
+        btn_add_customer = QPushButton('Add Customer', self)
+        btn_add_customer.clicked.connect(self.btn_add_customer_on_click)
 
+        btn_start_transaction = QPushButton('Start Transaction', self)
+        btn_start_transaction.clicked.connect(self.btn_start_transaction_on_click)
 
-grid = QGridLayout()
+        main_layout = QVBoxLayout()
+        main_layout.addWidget(btn_view_products)
+        main_layout.addWidget(btn_view_oos_products)
+        main_layout.addWidget(btn_view_recent_customers)
+        main_layout.addWidget(btn_add_customer)
+        main_layout.addWidget(btn_start_transaction)
 
-def clear_widgets():
-    for widget in widgets:
+        self.setLayout(main_layout)
 
-        if widgets[widget] != []:
-            widgets[widget][-1].hide()
+        self.show()
 
-        for i in range(0, len(widgets[widget])):
-            widgets[widget].pop()
+    def btn_view_products_on_click(self):
+        self.cams = ViewProductsWindow()
+        self.cams.show()
+        self.close()
 
-def homepage():
+    def btn_view_out_of_stock_products_on_click(self):
+        self.cams = ViewOOSProductsWindow()
+        self.cams.show()
+        self.close()
 
-    button_inventory = QPushButton("Inventory")
-    button_inventory.setCursor(QCursor(QtCore.Qt.PointingHandCursor))
-    button_inventory.setStyleSheet("border: 4px solid white;" + "color:white;")
-    button_inventory.clicked.connect(clear_widgets)
-    button_inventory.clicked.connect(inventory_page)
-    widgets["button_inventory"].append(button_inventory)
-    grid.addWidget(widgets["button_inventory"][-1], 0, 0)
+    def btn_view_recent_customers_on_click(self):
+        self.cams = ViewRecentCustomers()
+        self.cams.show()
+        self.close()
 
-    transaction = QPushButton("Transaction")
-    transaction.setCursor(QCursor(QtCore.Qt.PointingHandCursor))
-    transaction.setStyleSheet("border: 4px solid white;" + "color:white;")
-    transaction.clicked.connect(clear_widgets)
-    transaction.clicked.connect(point_of_sale_display)
-    widgets["transaction"].append(transaction)
+    def btn_add_customer_on_click(self):
+        self.cams = AddCustomerWindow()
+        self.cams.show()
+        self.close()
 
-    grid.addWidget(widgets["transaction"][-1], 1, 0)
-    transaction.clicked.connect(point_of_sale_display)
-
-    customer = QPushButton("Customer")
-    customer.setCursor(QCursor(QtCore.Qt.PointingHandCursor))
-    customer.setStyleSheet("border: 4px solid white;" + "color:white;")
-    customer.style()
-    customer.clicked.connect(clear_widgets)
-    customer.clicked.connect(customer_display)
-    widgets["customer"].append(customer)
-
-    grid.addWidget(widgets["customer"][-1], 2, 0)
-
-
-def inventory_page():
-
-    table = QTableWidget()
-    table.height()
-    table.width()
-    grid.addWidget(table, 1, 0)
-    table.setColumnCount(6)
-    data = controller.show_products()[1]
-    table.setHorizontalHeaderLabels(["product_id",
-                                     "vendor_id",
-                                     "product_name",
-                                     "product_description",
-                                     "inventory_stock",
-                                     "price"])
-    table.setStyleSheet("color:white;")
-    for row, end in enumerate(data):
-
-        data = QTableWidgetItem(end[0])
-        data.setTextAlignment(4)
-        table.insertRow(row)
-        table.setItem(row, 0, QTableWidgetItem(str(end[0])))
-        table.setItem(row, 1, QTableWidgetItem(str(end[1])))
-        table.setItem(row, 2, QTableWidgetItem(str(end[2])))
-        table.setItem(row, 3, QTableWidgetItem(str(end[3])))
-        table.setItem(row, 4, QTableWidgetItem(str(end[4])))
-        table.setItem(row, 5, QTableWidgetItem(str(end[5])))
-    add_to_inventory_button = QPushButton("Add to Inventory")
-    add_to_inventory_button.setCursor(QCursor(QtCore.Qt.PointingHandCursor))
-    grid.addWidget(add_to_inventory_button,0,0)
-    add_to_inventory_button.setStyleSheet("background-color: white;")
-    add_to_inventory_button.clicked.connect(customer_display)
-
-    def add_to_inventory_display():
-        window.setWindowTitle("Add to Inventory")
-
-        product_name_label = QLabel("Product Name:")
-        grid.addWidget(product_name_label,1,1)
-        product_name_label.setStyleSheet("")
+    def btn_start_transaction_on_click(self):
+        self.cams = StartTransactionWindow()
+        self.cams.show()
+        self.close()
 
 
-        # function that sets all the labels, buttons, line edits in place Window title and title, background color
-def customer_display():
-    window.setWindowTitle("Add Customer")
-    title = QLabel('Customer Information')
-    title.setStyleSheet("color:white;")
-    grid.addWidget(title,0,1)
-    # Label and input for customer_first_name
-    first_name_label = QLabel('First Name')
-    first_name_label.setStyleSheet("color: white;")
-    grid.addWidget(first_name_label,1,0)
-    first_name_input = QLineEdit()
-    first_name_input.setStyleSheet("background: white;")
-    grid.addWidget(first_name_input, 1, 1)
-    # Label and input for customer_last_name
-    last_name_label = QLabel('Last Name')
-    last_name_label.setStyleSheet("color: white;")
-    grid.addWidget(last_name_label, 2, 0)
-    last_name_input = QLineEdit()
-    last_name_input.setStyleSheet("background: white;")
-    grid.addWidget(last_name_input, 2, 1)
-    # Label and input for customer_phone
-    phone_label = QLabel('Phone Number')
-    phone_label.setStyleSheet("color: white;")
-    grid.addWidget(phone_label, 3, 0)
-    phone_input = QLineEdit()
-    phone_input.setStyleSheet("background: white;")
-    grid.addWidget(phone_input, 3, 1)
-        # Label and input for customer_email
-    email_label = QLabel('Email')
-    email_label.setStyleSheet("color: white;")
-    grid.addWidget(email_label, 4, 0)
-    email_input = QLineEdit()
-    email_input.setStyleSheet("background: white;")
-    grid.addWidget(email_input, 4, 1)
-        # Label and input for customer_address
-    address_label = QLabel('Address')
-    address_label.setStyleSheet("color: white;")
-    grid.addWidget(address_label, 5, 0)
-    address_input = QLineEdit()
-    address_input.setStyleSheet("background: white;")
-    grid.addWidget(address_input, 5, 1)
-        # Button adds info from gui inputs to customer_table when clicked
-    add_button = QPushButton('Add Customer')
-    add_button.setStyleSheet("background: white;")
-    grid.addWidget(add_button,6,1)
-        # Label that confirms wheather add has worked or not
-    feedback_label = QLabel()
-    feedback_label.setStyleSheet("color:white;")
-    grid.addWidget(feedback_label,7,1)
+class ViewProductsWindow(QWidget):
+    def __init__(self, parent=None):
+        super().__init__(parent)
+        self.setWindowTitle('View Products')
 
-    def add_new_customer():
-        # Function that connects add customer line edits and button to database
-        try:
-            customer_first_name = first_name_input.text()
-            customer_last_name = last_name_input.text()
-            customer_phone = phone_input.text()
-            customer_email = email_input.text()
-            customer_address = address_input.text()
-            to_controller = add_customer(customer_first_name,customer_last_name,
-                                         customer_phone,customer_email,customer_address)
-        except Exception as e:
-            feedback_label.setText(str(e))
+        self.cams = None
+        self.top = 100
+        self.left = 100
+        self.width = 680
+        self.height = 500
+        self.setGeometry(self.top, self.left, self.width, self.height)
 
+        # Return Button
+        self.pushButton = QPushButton(self)
+        self.pushButton.setText('Return to Homepage')
+        self.pushButton.setGeometry(260, 450, 150, 30)
+        self.pushButton.clicked.connect(self.show_homepage)
+
+        # create table view and set model
+        self.table_view = QTableView()
+        self.model = QStandardItemModel()
+        self.table_view.setModel(self.model)
+
+        cols = show_products()[0]
+
+        self.model.setHorizontalHeaderLabels(cols)
+        self.table_view.verticalHeader().setVisible(False)
+
+        for col in range(len(cols)):
+            self.table_view.setColumnWidth(col, 200)
+
+        rows = show_products()[1]
+
+        self.model.insertRows(0, len(rows), QModelIndex())
+
+        for row, data in enumerate(rows):
+            self.model.setItem(row, 0, QStandardItem(f"{rows[row][0]}"))
+            self.model.setItem(row, 1, QStandardItem(f"{rows[row][1]}"))
+            self.model.setItem(row, 2, QStandardItem(f"{rows[row][2]}"))
+            self.model.setItem(row, 3, QStandardItem(f"{rows[row][3]}"))
+            self.model.setItem(row, 4, QStandardItem(f"{rows[row][4]}"))
+            self.model.setItem(row, 5, QStandardItem(f"{rows[row][5]}"))
+
+            for col in range(len(cols)):
+                index = self.model.index(row, col, QModelIndex())
+                self.model.setData(index, Qt.AlignCenter, Qt.TextAlignmentRole)
+
+        # add table view to layout
+        layout = QVBoxLayout()
+        layout.addWidget(self.table_view)
+        layout.addWidget(self.pushButton)
+        self.setLayout(layout)
+
+    def show_homepage(self):
+        self.cams = HomepageWindow()
+        self.cams.show()
+        self.close()
+
+
+class ViewOOSProductsWindow(QWidget):
+    def __init__(self, parent=None):
+        super().__init__(parent)
+        self.setWindowTitle('View Out of Stock Products')
+
+        self.cams = None
+        self.top = 100
+        self.left = 100
+        self.width = 680
+        self.height = 500
+        self.setGeometry(self.top, self.left, self.width, self.height)
+
+        # Return Button
+        self.pushButton = QPushButton(self)
+        self.pushButton.setText('Return to Homepage')
+        self.pushButton.setGeometry(260, 450, 150, 30)
+        self.pushButton.clicked.connect(self.show_homepage)
+
+        # create table view and set model
+        self.table_view = QTableView()
+        self.model = QStandardItemModel()
+        self.table_view.setModel(self.model)
+
+        cols = show_out_of_stock_products()[0]
+
+        self.model.setHorizontalHeaderLabels(cols)
+        self.table_view.verticalHeader().setVisible(False)
+
+        for col in range(len(cols)):
+            self.table_view.setColumnWidth(col, 200)
+
+        rows = show_out_of_stock_products()[1]
+
+        self.model.insertRows(0, len(rows), QModelIndex())
+
+        for row, data in enumerate(rows):
+            self.model.setItem(row, 0, QStandardItem(f"{rows[row][0]}"))
+            self.model.setItem(row, 1, QStandardItem(f"{rows[row][1]}"))
+            self.model.setItem(row, 2, QStandardItem(f"{rows[row][2]}"))
+            self.model.setItem(row, 3, QStandardItem(f"{rows[row][3]}"))
+            self.model.setItem(row, 4, QStandardItem(f"{rows[row][4]}"))
+            self.model.setItem(row, 5, QStandardItem(f"{rows[row][5]}"))
+
+            for col in range(len(cols)):
+                index = self.model.index(row, col, QModelIndex())
+                self.model.setData(index, Qt.AlignCenter, Qt.TextAlignmentRole)
+
+        # add table view to layout
+        layout = QVBoxLayout()
+        layout.addWidget(self.table_view)
+        layout.addWidget(self.pushButton)
+        self.setLayout(layout)
+
+    def show_homepage(self):
+        self.cams = HomepageWindow()
+        self.cams.show()
+        self.close()
+
+
+class ViewRecentCustomers(QWidget):
+    def __init__(self, parent=None):
+        super().__init__(parent)
+        self.setWindowTitle('View Recent Customers')
+
+        self.cams = None
+        self.top = 100
+        self.left = 100
+        self.width = 680
+        self.height = 500
+        self.setGeometry(self.top, self.left, self.width, self.height)
+
+        # Return Button
+        self.pushButton = QPushButton(self)
+        self.pushButton.setText('Return to Homepage')
+        self.pushButton.setGeometry(260, 450, 150, 30)
+        self.pushButton.clicked.connect(self.show_homepage)
+
+        # create table view and set model
+        self.table_view = QTableView()
+        self.model = QStandardItemModel()
+        self.table_view.setModel(self.model)
+
+        cols = get_customers_last_month()[0]
+
+        self.model.setHorizontalHeaderLabels(cols)
+        self.table_view.verticalHeader().setVisible(False)
+
+        for col in range(len(cols)):
+            self.table_view.setColumnWidth(col, 200)
+
+        rows = get_customers_last_month()[1]
+
+        self.model.insertRows(0, len(rows), QModelIndex())
+
+        for row, data in enumerate(rows):
+            self.model.setItem(row, 0, QStandardItem(f"{rows[row][0]}"))
+            self.model.setItem(row, 1, QStandardItem(f"{rows[row][1]}"))
+            self.model.setItem(row, 2, QStandardItem(f"{rows[row][2]}"))
+            self.model.setItem(row, 3, QStandardItem(f"{rows[row][3]}"))
+
+            for col in range(len(cols)):
+                index = self.model.index(row, col, QModelIndex())
+                self.model.setData(index, Qt.AlignCenter, Qt.TextAlignmentRole)
+
+        # add table view to layout
+        layout = QVBoxLayout()
+        layout.addWidget(self.table_view)
+        layout.addWidget(self.pushButton)
+        self.setLayout(layout)
+
+    def show_homepage(self):
+        self.cams = HomepageWindow()
+        self.cams.show()
+        self.close()
+
+
+class AddCustomerWindow(QWidget):
+    def __init__(self, parent=None):
+        super().__init__(parent)
+        self.setWindowTitle('Add Customer')
+
+        self.cams = None
+        self.top = 100
+        self.left = 100
+        self.width = 680
+        self.height = 500
+        self.setGeometry(self.top, self.left, self.width, self.height)
+
+        # Return to home
+        self.homeButton = QPushButton(self)
+        self.homeButton.setText('Return to Homepage')
+        self.homeButton.setGeometry(260, 450, 150, 30)
+        self.homeButton.clicked.connect(self.show_homepage)
+
+        self.addCustomerButton = QPushButton(self)
+        self.addCustomerButton.setText('Add Customer')
+        self.addCustomerButton.setGeometry(260, 450, 150, 30)
+        self.addCustomerButton.clicked.connect(self.add_customer)
+
+        main_layout = QVBoxLayout()
+
+        self.cus_first_name = QLineEdit("", self)
+        self.cus_last_name = QLineEdit("", self)
+        self.cus_phone = QLineEdit("", self)
+        self.cus_email = QLineEdit("", self)
+        self.cus_address = QLineEdit("", self)
+
+        first_name_row = QHBoxLayout()
+        first_name_row.addWidget(QLabel("First Name: "))
+        first_name_row.addWidget(self.cus_first_name)
+        main_layout.addLayout(first_name_row)
+
+        last_name_row = QHBoxLayout()
+        last_name_row.addWidget(QLabel("Last Name: "))
+        last_name_row.addWidget(self.cus_last_name)
+        main_layout.addLayout(last_name_row)
+
+        phone_row = QHBoxLayout()
+        phone_row.addWidget(QLabel("Phone: "))
+        phone_row.addWidget(self.cus_phone)
+        main_layout.addLayout(phone_row)
+
+        email_row = QHBoxLayout()
+        email_row.addWidget(QLabel("Email: "))
+        email_row.addWidget(self.cus_email)
+        main_layout.addLayout(email_row)
+
+        address_row = QHBoxLayout()
+        address_row.addWidget(QLabel("Address: "))
+        address_row.addWidget(self.cus_address)
+        main_layout.addLayout(address_row)
+
+        main_layout.addWidget(self.addCustomerButton)
+        main_layout.addWidget(self.homeButton)
+        self.setLayout(main_layout)
+
+    def show_homepage(self):
+        self.cams = HomepageWindow()
+        self.cams.show()
+        self.close()
+
+    def add_customer(self):
+        add_customer(self.cus_first_name.text(), self.cus_last_name.text(), self.cus_phone.text(),
+                     self.cus_email.text(),
+                     self.cus_address.text())
+
+
+class StartTransactionWindow(QWidget):
+    def __init__(self, parent=None):
+        super().__init__(parent)
+        self.setWindowTitle('Start Transaction')
+
+        self.cams = None
+        self.top = 100
+        self.left = 100
+        self.width = 680
+        self.height = 500
+        self.setGeometry(self.top, self.left, self.width, self.height)
+
+        self.current_invoice_id = None
+        self.current_customer_id = None
+
+        # Home Button
+        self.home_button = QPushButton(self)
+        self.home_button.setText('Return to Homepage')
+        self.home_button.clicked.connect(self.show_homepage)
+
+        main_layout = QVBoxLayout()
+
+        self.cus_cbo = QComboBox()
+        self.cus_button = QPushButton("Select Customer")
+        self.cus_button.clicked.connect(self.select_customer_on_click)
+
+        self.cus_cbo.addItem('')
+        row = get_customer_names()
+        for i in row[1]:
+            data = str(i[0])
+            self.cus_cbo.addItem(data)
+
+        choose_customer_row = QHBoxLayout()
+        choose_customer_row.addWidget(QLabel("Customer: "))
+        choose_customer_row.addWidget(self.cus_cbo)
+        choose_customer_row.addWidget(self.cus_button)
+
+        self.prod_cbo = QComboBox()
+        self.prod_button = QPushButton("Add Product ")
+        self.prod_button.clicked.connect(self.select_product_on_click)
+        self.prod_quantity_line = QLineEdit()
+        self.prod_quantity_line.setFixedWidth(50)
+
+        self.prod_cbo.setEnabled(False)
+        self.prod_button.setEnabled(False)
+        self.prod_quantity_line.setEnabled(False)
+
+        self.prod_cbo.addItem('')
+        row = get_product_names()
+        for i in row[1]:
+            data = str(i[0])
+            self.prod_cbo.addItem(data)
+
+        choose_product_row = QHBoxLayout()
+        choose_product_row.addWidget(QLabel("Product: "))
+        choose_product_row.addWidget(self.prod_cbo)
+        choose_product_row.addWidget(self.prod_quantity_line)
+        choose_product_row.addWidget(self.prod_button)
+
+        self.btn_complete_invoice = QPushButton(self)
+        self.btn_complete_invoice.setText('Complete Invoice')
+        self.btn_complete_invoice.clicked.connect(self.btn_complete_invoice_on_click)
+        self.btn_complete_invoice.setEnabled(False)
+
+        self.btn_show_invoice = QPushButton(self)
+        self.btn_show_invoice.setText('Show Invoice')
+        self.btn_show_invoice.clicked.connect(self.btn_show_invoice_on_click)
+        self.btn_show_invoice.setEnabled(False)
+
+        main_layout.addLayout(choose_customer_row)
+        main_layout.addLayout(choose_product_row)
+        main_layout.addWidget(self.btn_complete_invoice)
+        main_layout.addWidget(self.btn_show_invoice)
+        main_layout.addWidget(self.home_button)
+
+        self.setLayout(main_layout)
+
+    def show_homepage(self):
+        self.cams = HomepageWindow()
+        self.cams.show()
+        self.close()
+
+    def select_customer_on_click(self):
+        self.prod_cbo.setEnabled(True)
+        self.prod_button.setEnabled(True)
+        self.prod_quantity_line.setEnabled(True)
+        self.btn_complete_invoice.setEnabled(True)
+        cus_id = get_customer_by_name(self.cus_cbo.currentText())[1][0][0]
+        self.current_customer_id = cus_id
+        add_invoice(cus_id, 'NOW()')
+        self.current_invoice_id = get_most_recent_invoice_id_by_date()
+        self.cus_cbo.setEnabled(False)
+        self.cus_button.setEnabled(False)
+
+    def select_product_on_click(self):
+        prod_id = get_product_id(self.prod_cbo.currentText())
+        if self.prod_quantity_line.text() == '':
+            self.prod_quantity_line.setText('1')
+        if check_if_invoice_product_exists(self.current_invoice_id, prod_id):
+            increase_invoice_product_inventory(self.prod_quantity_line.text(), self.current_invoice_id, prod_id)
         else:
-            if to_controller == 1:
-                feedback_label.setText("Customer Added")
+            current_inventory_count = get_product_quantity_by_id(prod_id)[0][1]
+            print(current_inventory_count)
+            if current_inventory_count < int(self.prod_quantity_line.text()):
+                print('Product oos!')
             else:
-                feedback_label.setText("Failed to Add Customer")
+                add_product_to_invoice_products(self.current_invoice_id, prod_id, self.prod_quantity_line.text())
+                decrease_inventory(self.prod_quantity_line.text(), prod_id)
 
-    add_button.clicked.connect(add_new_customer)
+    def btn_complete_invoice_on_click(self):
+        self.btn_complete_invoice.setEnabled(False)
+        self.prod_cbo.setEnabled(False)
+        self.prod_button.setEnabled(False)
+        self.prod_quantity_line.setEnabled(False)
+        self.btn_show_invoice.setEnabled(True)
 
-def point_of_sale_display():
-    # Function that sets all labels, and combo boxes in place
-
-    window.setWindowTitle("Point of Sale")
-    select_customer_label = QLabel('Select Customer:')
-    grid.addWidget(select_customer_label,0,0)
-    select_customer_label.setStyleSheet("color: white;")
-    # Combo box which allows selection from list of customers from customers table in pgp database
-    customer_combobox = QComboBox(window)
-    grid.addWidget(customer_combobox,0,1)
-    customer_combobox.setStyleSheet("background-color: white;")
-    # Uses for loop to add customer_fullname to combo box
-    customer_combobox.addItem('')
-    row = get_customer_names()
-    for i in row[1]:
-        data = str(i[0])
-        customer_combobox.addItem(data)
-
-    show_invoice_button = QPushButton('Show Invoice')  # Button when clicked shows invoice
-    grid.addWidget(show_invoice_button, 4, 1)
-    show_invoice_button.setStyleSheet("background-color: white;")
-
-    # Label that holds customer_id, hidden from user, for use by later function
-    customer_id_label = QLabel("")
-    grid.addWidget(customer_id_label,0,3)
-
-    # function which updates customer_id label with corresponding id for customer name selected
-    def update_id_label():
-
-        selected_name = customer_combobox.currentText()
-        customer_fullname_split = selected_name.split(' ')
-        customer_first_name = customer_fullname_split[0]
-        customer_last_name = customer_fullname_split[1]
-        info = get_customer_id_by_customer_name(customer_first_name,customer_last_name)
-
-        customer_id_label.setText(info)
-
-    customer_combobox.currentIndexChanged.connect(update_id_label)
-        # Label for select product combobox
-    select_product_label = QLabel('Select Product:')
-    grid.addWidget(select_product_label,1,0)
-    select_product_label.setStyleSheet("color: white;")
-        # Label that displays price of item selected
-    price_label = QLabel()
-    grid.addWidget(price_label,1,3)
-    price_label.setStyleSheet('color: white;')
-        # Combo box which allows selection from list of products from products table in pgp database
-    product_combobox = QComboBox(window)
-    grid.addWidget(product_combobox, 1, 1)
-    product_combobox.setStyleSheet("background-color: white;")
-    product_combobox.addItem('')
-
-    quantity_label = QLabel('Qty:')
-    grid.addWidget(quantity_label,1,2)
-    quantity_label.setStyleSheet("color: white;"+"max-width: 20px")
-
-    quantity_line_edit = QLineEdit()
-    grid.addWidget(quantity_line_edit,1,3)
-    quantity_line_edit.setStyleSheet("background-color: white;"+"max-width: 50px")
-
-        # Uses for loop to add product_name from product table
-
-    row = get_product_names()
-    for i in row[1]:
-        pdata = str(i[0])
-        product_combobox.addItem(pdata)
-
-    # Function which updates price label with price of current product selection from combobox
-    def update_price_label():
-
-        product_name = product_combobox.currentText()
-        info = get_price_with_product_name(product_name)
-        price = str(f"List Price: ${info[11:16]} ea")
-        price_label.setText(price)
+    def btn_show_invoice_on_click(self):
+        self.cams = ViewCurrentInvoiceWindow(self.current_invoice_id)
+        self.cams.show()
 
 
-    product_combobox.currentIndexChanged.connect(update_price_label)
-    # Button which when clicked adds the customer id to the invoices table
-    add_customer_id_to_invoices_button = QPushButton("Add Customer")
-    grid.addWidget(add_customer_id_to_invoices_button, 0, 4)
-    add_customer_id_to_invoices_button.setStyleSheet("background-color: white;")
-    # Label which holds the invoice id
-    invoice_id_label = QLabel()
-    grid.addWidget(invoice_id_label, 4, 0)
-    invoice_id_label.setStyleSheet('color: white;')
-    # Label which holds product_id
-    product_id_label = QLabel()
-    grid.addWidget(product_id_label, 4, 0)
-    product_id_label.setStyleSheet('color: white;')
-    # Button which when clicked adds product id to invoice products table
-    add_product_id_to_invoice_products_button = QPushButton("Add Product")
-    grid.addWidget(add_product_id_to_invoice_products_button, 1, 4)
-    add_product_id_to_invoice_products_button.setStyleSheet("background-color: white;")
-    # Label that carries either confirmation of item added or an error code
-    feedback_label = QLabel()
-    feedback_label.setStyleSheet("color:white;")
-    grid.addWidget(feedback_label, 3, 2)
+class ViewCurrentInvoiceWindow(QWidget):
+    def __init__(self, current_invoice, parent=None):
+        super().__init__(parent)
+        self.setWindowTitle('View Current Invoice')
 
-    # Function when called adds the selected customer id to invoices table
-    # and gets invoice id from this entry into invoices table for later use
-    def add_customer_to_invoices():
-        try:
-            customer_id = customer_id_label.text()
-            id_to_controller = add_customer_id_to_invoices(int(customer_id))
-            invoice_id = get_most_recent_invoice_id_by_date()
+        self.cams = None
+        self.top = 100
+        self.left = 100
+        self.width = 680
+        self.height = 500
+        self.setGeometry(self.top, self.left, self.width, self.height)
 
-        except Exception as e:
-            feedback_label.setText(str(e))
+        self.current_invoice_id = current_invoice
 
-        else:
-            if id_to_controller == 1:
-                feedback_label.setText("Successfully Added")
-            else:
-                feedback_label.setText("Failed to Add")
+        # Return Button
+        self.pushButton = QPushButton(self)
+        self.pushButton.setText('Close')
+        self.pushButton.setGeometry(260, 450, 150, 30)
+        self.pushButton.clicked.connect(self.show_homepage)
 
-        # Function which gets product id from the products table using the product name selected from combobox
-        # and uses previous query of invoice id then inserts both of them into invoice_products
-        def add_product_to_invoice():
-            try:
-                product_choice = product_combobox.currentText()
-                quantity = quantity_line_edit.text()
+        # create table view and set model
+        self.table_view = QTableView()
+        self.model = QStandardItemModel()
+        self.table_view.setModel(self.model)
 
-                product_id = get_product_id(product_choice)
-                product_info_to_controller = add_product_to_invoice_products(invoice_id,product_id,quantity)
+        cols = show_current_invoice(self.current_invoice_id)[0]
 
-                # SHOULD BE A COUNT OF HOW MANY TIMES A PRODUCT ID WITH A SPECIFIED INVOICE ID IS PRESENT FOR
-                # QUANTITY PURPOSES
+        self.model.setHorizontalHeaderLabels(cols)
 
-            except Exception as e:
-                feedback_label.setText(str(e))
-            else:
-                if product_info_to_controller == 1:
-                    feedback_label.setText("Successfully Added")
-                else:
-                    feedback_label.setText("Failed to Add")
+        self.table_view.verticalHeader().setVisible(False)
 
-        add_product_id_to_invoice_products_button.clicked.connect(add_product_to_invoice)
-    add_customer_id_to_invoices_button.clicked.connect(add_customer_to_invoices)
+        for col in range(len(cols)):
+            self.table_view.setColumnWidth(col, 200)
+
+        rows = show_current_invoice(self.current_invoice_id)[1]
+
+        self.model.insertRows(0, len(rows), QModelIndex())
+
+        for row, data in enumerate(rows):
+            self.model.setItem(row, 0, QStandardItem(f"{rows[row][0]}"))
+            self.model.setItem(row, 1, QStandardItem(f"{rows[row][1]}"))
+            self.model.setItem(row, 2, QStandardItem(f"{rows[row][2]}"))
+            self.model.setItem(row, 3, QStandardItem(f"{rows[row][3]}"))
+            self.model.setItem(row, 4, QStandardItem(f"{rows[row][4]}"))
+
+            for col in range(len(cols)):
+                index = self.model.index(row, col, QModelIndex())
+                self.model.setData(index, Qt.AlignCenter, Qt.TextAlignmentRole)
+
+        # add table view to layout
+        layout = QVBoxLayout()
+        layout.addWidget(self.table_view)
+        layout.addWidget(self.pushButton)
+        self.setLayout(layout)
+
+    def show_homepage(self):
+        self.close()
 
 
-def invoice_display():
-    invoice_window.show()
-    grid2 = QGridLayout(invoice_window)
-    invoice_window.setWindowTitle("Invoice")
-    invoice_title = QLabel('Purchase Information')
-    invoice_title.setStyleSheet("color:white;"+"font-size: 20px;")
-    grid2.addWidget(invoice_title, 1, 0)
-
-    table = QTableWidget()
-    table.setStyleSheet("background-color: white;")
-    table.height()
-    table.width()
-    grid2.addWidget(table, 2, 0)
-    table.setColumnCount(6)
-    data = controller.show_products()[1]
-    for row, end in enumerate(data):
-
-        idDato = QTableWidgetItem(end[0])
-        idDato.setTextAlignment(4)
-        table.insertRow(row)
-        table.setItem(row, 0, idDato)
-        table.setItem(row, 1, QTableWidgetItem(str(end[1])))
-        table.setItem(row, 2, QTableWidgetItem(str(end[2])))
-        table.setItem(row, 3, QTableWidgetItem(str(end[3])))
-        table.setItem(row, 4, QTableWidgetItem(str(end[4])))
-        table.setItem(row, 5, QTableWidgetItem(str(end[5])))
-
-homepage()
-#point_of_sale_display()
-window.setLayout(grid)
-
-window.show()
-
-sys.exit(app.exec())
+if __name__ == '__main__':
+    app = QApplication(sys.argv)
+    ex = HomepageWindow()
+    sys.exit(app.exec_())
